changed reset button values to percents

+++ src/com/runnershigh/RHDrawable.java
@@ -6,7 +6,7 @@ public class RHDrawable extends Mesh {
 	protected float width;
 	protected float height;
 	
-	public RHDrawable(int _x, int _y, int _z, int _width, int _height) {
+	public RHDrawable(float _x, float _y, float _z, float _width, float _height) {
 		x = _x;
 		y = _y;
 		z = _z;
